Move GetSegments to AnsiConsoleExtensions
Swap the parameters to make the "active" type (in this case the console)
the type that is extended and the "passive" renderable the parameter.
