+++ src/Spectre.Console/Internal/Backends/Ansi/AnsiConsoleBackend.cs
@@ -33,7 +33,7 @@ namespace Spectre.Console
         public void Write(IRenderable renderable)
         {
             var builder = new StringBuilder();
-            foreach (var segment in renderable.GetSegments(_console))
+            foreach (var segment in _console.GetSegments(renderable))
             {
                 if (segment.IsControlCode)
                 {
